forward and backward substitutions for block triangular matrices and Ensembles, tested
