+++ eggshell/constants.h
@@ -9,6 +9,6 @@ const Eigen::Vector3d kGravity(0, 0, -9.8);  // in m/s^2
 
 // A matrix is considered well conditioned when the condition number is below
 // this value.
-constexpr double kGoodConditionNumber = 1e8;
+constexpr double kGoodConditionNumber = 1e7;
 
 }  // namespace
